+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changes that are not related to specific components
 #### Fixed
 
 - [Component] What bugs/typos are fixed?
+- [FileInput] FileInput accepts capitalized extensions (.png vs .PNG)
 
 ### Core